Move CheckURLFree and CloneURL to sandbox
