+++ pkg/tools/grpcutils/url.go
@@ -93,24 +93,3 @@ func TargetToNetAddr(target string) (network, addr string) {
 
 	return network, target
 }
-
-// CheckURLFree returns is given url is free for Listen
-func CheckURLFree(u *url.URL) bool {
-	ln, err := net.Listen(TargetToNetAddr(URLToTarget(u)))
-	if err == nil {
-		err = ln.Close()
-	}
-	return err == nil
-}
-
-// CloneURL clones given url
-func CloneURL(u *url.URL) *url.URL {
-	if u == nil {
-		return nil
-	}
-
-	cloned := new(url.URL)
-	*cloned = *u
-
-	return cloned
-}
